Factor ModuleContext out of translate.rs into new module.
Initial work on refactoring unwieldly translate.rs.
(depends on earlier small patch https://github.com/solana-labs/move/pull/245)
